Rename gen module to generate
