+++ crates/aide/src/generate.rs
