+++ examples/example-axum/src/docs.rs
@@ -20,7 +20,7 @@ pub fn docs_routes(state: AppState) -> ApiRouter {
     // As a result, the `serve_redoc` route will
     // have the `text/html` content-type correctly set
     // with a 200 status.
-    aide::gen::infer_responses(true);
+    aide::generate::infer_responses(true);
 
     let router: ApiRouter = ApiRouter::new()
         .api_route_with(
@@ -58,7 +58,7 @@ pub fn docs_routes(state: AppState) -> ApiRouter {
 
     // Afterwards we disable response inference because
     // it might be incorrect for other routes.
-    aide::gen::infer_responses(false);
+    aide::generate::infer_responses(false);
 
     router
 }
